Add rotation

+++ src/main/kotlin/chart/difficulty/_obstacles.kt
@@ -5,10 +5,10 @@ package chart.difficulty
 import com.google.gson.annotations.SerializedName
 
 data class _obstacles (
-    @SerializedName("_time") var _time : Float,
+    @SerializedName("_time") var _time : Double,
     @SerializedName("_lineIndex") val _lineIndex : Int,
     @SerializedName("_type") val _type : Int,
-    @SerializedName("_duration") var _duration : Float,
+    @SerializedName("_duration") var _duration : Double,
     @SerializedName("_width") val _width : Int,
     @SerializedName("_customData") val _obstacleCustomData : _obstacleCustomData?
 )
